Be less aggressive about absorption of panics
pq uses panic() to raise certain errors, but it seems to me that a
couple of the cases in this switch are overly aggressive about
absorbing panics, leading to confusing diagnoses.

Unfortunately, this change is probably pretty sensitive and going to
blow up a few things previously absorbed, but it appears to be
definitely hiding bugs and there's no great way to start getting them
treated.

The case that finally instigated this was panic-ing in mustParse
yielding cryptic failures.

Signed-off-by: Daniel Farina <daniel@fdr.io>